requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
click==8.3.0
croniter==6.0.0
distro==1.9.0
fastapi==0.119.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
loguru==0.7.3
openai==2.6.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==7.0.1
rq==2.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
tavily-python==0.5.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
